Rust: migrated json to use serde.
