+++ json/json.rs/Cargo.toml
@@ -1,5 +1,10 @@
 [package]
 name = "json-rs"
 version = "0.0.1"
+
 [dependencies]
-rustc-serialize = "*"
+serde = "*"
+serde_macros = "*"
+
+[profile.release]
+lto = true
